requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
Pillow==10.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
blinker==2.1.2
dlib==19.24.2
cmake==3.27.7
scikit-image==0.21.0
scipy==1.11.3
matplotlib==3.7.2
imageio==2.31.5
networkx==3.1
tifffile==2023.7.10
packaging==23.1
pydantic==2.4.2
pydantic-core==2.10.1
annotated-types==0.4.0
typing-extensions==4.7.1
